add check if process group is the same a stdout process group
when the current process group is not the same as the processgroup of
stdout, we have sent the process to the background with &. in this case
we have a tty but stdin and stdout are no longer interactive. That means
we have to call docker or docker compose non interactive to avoid
errors.

proper fix for #5
fixes #6

Signed-off-by: BlackEagle <ike.devolder@gmail.com>