run composer with or without running application
When there is a running application container found, the composer
container will be started in the same pid and network namespace as the
'first' found running application container. This will allow us to have
post-process hooks in composer that interact with the working
application without throwing a ton of errors.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>